requirements.txt
# Core dependencies
semantic-kernel==1.7.0
azure-ai-foundry>=1.0.0
azure-identity>=1.15.0
azure-ai-inference>=1.0.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI and logging
click>=8.1.0
rich>=13.7.0
structlog>=23.2.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0